requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.4
pytest==8.1.1
pyserial==3.5
python-multipart>=0.0.6
